# dist/swagger.yaml
swagger: '2.0'
info:
  version: 1.0.0
  title: Swagger btrshop-cloud
  contact:
    name: Nicolas D.
    url: https://github.com/Oupsla
schemes:
  - https
  - http
consumes:
  - application/json
produces:
  - application/json
paths:
...
  '/products/{ean}':
    get:
      summary: Info for a specific product
      operationId: showProductByEan
      tags:
        - products
      parameters:
        - name: ean
          in: path
          required: true
          description: The ean of the product to retrieve
          type: string
      responses:
        '200':
          description: Excected response to a valid request
          schema:
            $ref: '#/definitions/Product'
        default:
          description: unexpected error
          schema:
            $ref: '#/definitions/Error'
    post:
      summary: Post a new position
      operationId: addPosition
      tags:
        - products
      parameters:
        - name: ean
          in: path
          required: true
          description: The ean of the product
          type: string
        - name: position
          in: body
          description: Position to be added to the product
          required: true
          schema:
            $ref: '#/definitions/Position'
      responses:
        '200':
          description: product response
          schema:
            $ref: '#/definitions/Product'
        default:
          description: unexpected error
          schema:
            $ref: '#/definitions/Error'
definitions:
  Product:
    required:
      - name
      - ean
      - description
    properties:
      brand:
        type: string
      category:
        type: string
      color:
        type: string
      description:
        type: string
      depth:
        $ref: '#/definitions/QuantitativeValue'
      ean:
        type: string
      height:
        $ref: '#/definitions/QuantitativeValue'
      logo:
        type: string
      offers:
        type: array
        items:
          $ref: '#/definitions/Offer'
      model:
        type: string
      name:
        type: string
      weight:
        $ref: '#/definitions/QuantitativeValue'
      width:
        $ref: '#/definitions/QuantitativeValue'
  QuantitativeValue:
    required:
      - unitText
      - value
    properties:
      minValue:
        type: number
      maxValue:
        type: number
      unitCode:
        type: string
      unitText:
        type: string
      value:
        type: number
  Offer:
    required:
      - price
      - priceCurrency
    properties:
      price:
        type: number
      priceCurrency:
        type: string
      validFrom:
        type: string
        format: date
      validThrough:
        type: string
        format: date
  Position:
    required:
      - lat
      - lng
    properties:
      lat:
        type: number
      lng:
        type: number
  Error:
    required:
      - code
      - message
    properties:
      code:
        type: integer
        format: int32
      message:
        type: string